refactor: Service 간의 구체 클래스 의존성 해결을 위한 인터페이스 생성
